refactor(party): split into helpers.py, views.py, and cog
Closes #166

- helpers.py: pure functions (parse_settings_text, parse_roles_from_text,
  validate_roles, parse_scheduled_time, _parse_roles_from_args,
  parse_allow_multiple, _parse_bool_value, format_timestamp) + constants
- views.py: all discord.ui.Modal and discord.ui.View classes
- party.py: Party(commands.Cog) only, imports from the above modules

Consolidates two inline _fmt_ts local functions into format_timestamp().
Replaces inline role-parsing in party_create with _parse_roles_from_args().
No behaviour change.
